pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mosaic-gpt"
version = "0.1.0"
description = "MOSAIC-GPT: Modular Open-Source Architecture for Intelligent Computation"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Lucas Krause", email = "lucas.krause@colostate.edu"},
]

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "datasets>=2.16.0",
    "pyyaml>=6.0",
    "wandb>=0.16.0",
    "tqdm>=4.66.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
include = ["mosaic*"]
